add support for resizable ring
Allows the ring to be resized as a cluster operation similar to join/leave.
Expansion is only allowed, but the restriction is temprorary, and shrinking
requires only small modifications.

When the operation is staged/commited the claimant calulates a resized ring,
running it through claim. It then schedules a "resize operation" for each
partition, via the next list. Each partition is scheduled for the operation because
during resize, each index potentially has data to send to others. Per partition,
the resize operation consists of several "resize transfers", each of which must be
completed to complete the entire operation. The first of these resize transfers is
scheduled by the claimant. Subsequent transfers are scheduled by the vnode that
owns the partition. When all of the transfers for a partition of have completed,
the operation is marked as such in the next list. Once the operation has completed
for *all* partitions the claimant installs the resized ring.

The handoff subsystem has been modified to add a new transfer type: resize_transfer.
In addition, the handoff sender now supports applying a function to keys that are
not sent (due to the handoff filter). The function operates on the unsent key and
a per-sender defined accumulator. Resize transfers are triggered like hinted and
ownership handoff by the vnode manager and vnode. The primary difference between
those transfers and a resize_transfer is that the latter includes a filter
that only sends keys between the source and target partitions if, for both, the
key hashes to the same position in the preflist. Resize transfers also include an
unsent keys function and accumulator that determines which partitions the key
would be destined for. The vnode uses this information to schedule its subsequent
resize transfers. A partition which has not been transferred to yet and is in the
returned list will be scheduled. No more resize transfers will be schedueld when
the returned list contains no partitions or only partitions that have already been
scheduled.

One major difference between typical transfers and a resize operation is the partition
does not delete its data (or unregister) after completing. For some partitions, they will
remain running but a portion of their data should be removed. Other partitions will no
longer live on their original owner should have all of their data deleted. This is
not addressed in this commit.

During resizing, how vnodes forward has been modified, as well. Unlike regular handoff
forwarding is performed only explicitly (when the vnode returns {forward,...} from
handle_handoff_command). The difference is that since the operation consists of multiple
transfers, and will support the ability to abort the operation, all requests are handled
locally and may be forwarded explicitly if they have a defined "request hash". To determine
the hash of the request, a new vnode callback has been added: request_hash. This callback is
optional, unless the application supports ring resizing. The callback can also return undefined
instead of a hash. In which case, the message will be always be handled locally only. Special care
must be taken to handle messages on a vnode forwarded from another, because primaries take part
in forwarding during resize. Coverage requests are not forwarded during a resize.